# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Soundwave Music Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /artists:
    get:
      tags: [Artists]
      operationId: listArtists
      summary: List all artists
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Filter by artist name
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of artists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artist"
  /artists/{id}:
    get:
      tags: [Artists]
      operationId: getArtist
      summary: Get artist profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artist profile with albums
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistDetail"
        "404":
          description: Artist not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Artist:
      type: object
      required: [id, name, albumCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        bio:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        albumCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ArtistDetail:
      type: object
      required: [id, name, albumCount, albums, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        bio:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isVerified:
          type: boolean
        albumCount:
          type: integer
        albums:
          type: array
          items:
            $ref: "#/components/schemas/Album"
        createdAt:
          type: string
          format: date-time
    Album:
      type: object
      required: [id, title, artistId, artistName, uploaderUsername, songCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        artistId:
          type: integer
        artistName:
          type: string
        coverUrl:
          type: ["string", "null"]
        uploaderUsername:
          type: string
        uploaderClerkId:
          type: string
        songCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...