# lib/api-zod/src/generated/types/artistDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Soundwave Music Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Album } from './album';

export interface ArtistDetail {
  id: number;
  name: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isVerified?: boolean;
  albumCount: number;
  albums: Album[];
  createdAt: Date;
}